fix: Close recipe modal when user query new valid ingredients
